Move "SyncMode" module into "restore" and rename to "RestoreOptions"
