+++ app/src/main/java/io/horizontalsystems/bankwallet/core/utils/RequestCode.kt
@@ -1,9 +1,9 @@
 package io.horizontalsystems.bankwallet.core.utils
 
 object ModuleCode {
-    const val SYNC_MODE = 1
+    const val RESTORE_EOS = 1
     const val RESTORE_WORDS = 2
-    const val RESTORE_EOS = 3
+    const val RESTORE_OPTIONS = 3
     const val BACKUP_WORDS = 4
     const val BACKUP_EOS = 5
     const val UNLOCK_PIN = 6
