+++ app/src/main/res/layout/activity_restore_options.xml
@@ -1,10 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
+    android:layout_height="match_parent">
 
     <io.horizontalsystems.bankwallet.ui.extensions.ShadowlessToolbarView
         android:id="@+id/shadowlessToolbar"
@@ -12,29 +10,25 @@
         android:layout_height="wrap_content"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        />
+        app:layout_constraintTop_toTopOf="parent" />
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/shadowlessToolbar"
-        >
+        app:layout_constraintTop_toBottomOf="@id/shadowlessToolbar">
 
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            >
+            android:orientation="vertical">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/fastSync"
                 android:layout_width="match_parent"
                 android:layout_height="60dp"
                 android:layout_marginTop="20dp"
-                android:background="@drawable/clickable_setting_item_bg"
-                >
+                android:background="@drawable/clickable_setting_item_bg">
 
                 <View
                     android:id="@+id/fastTopShade"
@@ -43,8 +37,7 @@
                     android:background="@color/steel_15"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    />
+                    app:layout_constraintTop_toTopOf="parent" />
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="0dp"
@@ -53,8 +46,7 @@
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toStartOf="@+id/fastCheckmarkIcon"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    >
+                    app:layout_constraintTop_toTopOf="parent">
 
                     <TextView
                         android:id="@+id/fastTitle"
@@ -76,6 +68,7 @@
                         android:textColor="@color/grey"
                         app:layout_constraintStart_toStartOf="@+id/fastTitle"
                         app:layout_constraintTop_toBottomOf="@+id/fastTitle" />
+
                 </androidx.constraintlayout.widget.ConstraintLayout>
 
                 <ImageView
@@ -86,16 +79,14 @@
                     android:src="@drawable/checkmark_orange"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    />
+                    app:layout_constraintTop_toTopOf="parent" />
 
                 <View
                     android:id="@+id/fastBottomShade"
                     android:layout_width="match_parent"
                     android:layout_height="0.5dp"
                     android:background="?SettingsSectionBottomShadeColor"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    />
+                    app:layout_constraintBottom_toBottomOf="parent" />
 
             </androidx.constraintlayout.widget.ConstraintLayout>
 
@@ -114,8 +105,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="60dp"
                 android:layout_marginTop="32dp"
-                android:background="@drawable/clickable_setting_item_bg"
-                >
+                android:background="@drawable/clickable_setting_item_bg">
 
                 <View
                     android:id="@+id/slowTopShade"
@@ -124,8 +114,7 @@
                     android:background="@color/steel_15"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    />
+                    app:layout_constraintTop_toTopOf="parent" />
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:layout_width="0dp"
@@ -134,8 +123,7 @@
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toStartOf="@+id/slowCheckmarkIcon"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    >
+                    app:layout_constraintTop_toTopOf="parent">
 
                     <TextView
                         android:id="@+id/slowTitle"
@@ -157,6 +145,7 @@
                         android:textColor="@color/grey"
                         app:layout_constraintStart_toStartOf="@+id/slowTitle"
                         app:layout_constraintTop_toBottomOf="@+id/slowTitle" />
+
                 </androidx.constraintlayout.widget.ConstraintLayout>
 
                 <ImageView
@@ -167,16 +156,14 @@
                     android:src="@drawable/checkmark_orange"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    />
+                    app:layout_constraintTop_toTopOf="parent" />
 
                 <View
                     android:id="@+id/slowBottomShade"
                     android:layout_width="match_parent"
                     android:layout_height="0.5dp"
                     android:background="?SettingsSectionBottomShadeColor"
-                    app:layout_constraintBottom_toBottomOf="parent"
-                    />
+                    app:layout_constraintBottom_toBottomOf="parent" />
 
             </androidx.constraintlayout.widget.ConstraintLayout>
 
@@ -191,6 +178,7 @@
                 android:textColor="@color/grey" />
 
         </LinearLayout>
+
     </androidx.core.widget.NestedScrollView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
